// src/app/models/compagnie.model.ts
export class CompagniePartenaire{
    idCompagnie?: number;
    libelle?: string;
    email?: string;
    numero?: number;
    adresse?: string;
    ninia?: string;
    rccm?: string;
    livraison?: JSON;
    created_by?: JSON;
    created_at?: Date;
    modified_by?: JSON;
    modified_at?: Date;
  }